+++ packages/components/bitcoin/src/blocks-queue/blocks-loader/blocks-loader.service.ts
@@ -86,7 +86,7 @@ export class BlocksQueueLoaderService implements OnModuleDestroy {
     const name = this.config.queueLoaderStrategyName;
 
     switch (name) {
-      case StrategyNames.PULL_NETWORK_PROVIDER:
+      case StrategyNames.PULL:
         this._loadingStrategy = new PullNetworkProviderStrategy(this.log, this.blockchainProviderService, queue, {
           maxRequestBlocksBatchSize: this.config.queueLoaderRequestBlocksBatchSize,
           concurrency: this.config.queueLoaderConcurrency,